crawler/requirements.txt
cryptography==37.0.4
wheel==0.37.1
Scrapy==2.5.0
scrapyd==1.2.1
scrapyd-client==1.2.2
python-scrapyd-api==2.1.2
scrapy-selenium==0.0.7
selenium==3.141.0
peewee==3.14.4
pandas==1.4.0
psycopg2-binary==2.9.3
celery==5.2.7
requests==2.23.0